refactor: group server media requests
