+++ apps/backend/src/server/mediaRequests/directImageIngestionRouting.ts
